Cambios 04/12/25 16:00

+++ Almacen STLCC/Models/Auditoria/Auditoria.cs
@@ -27,7 +27,7 @@ namespace Almacen_STLCC.Models.Auditoria
 
         [Required]
         [Column("id_registro")]
-        public long Id_Registro { get; set; }
+        public int Id_Registro { get; set; }
 
         [Column("descripcion", TypeName = "TEXT")]
         public string? Descripcion { get; set; }
